fix(discussions): reply count showing 01/011/0111 instead of 1/2/3
pg returns COUNT(*) as a string (bigint), so reply_count + 1 was string
concatenation instead of addition — '0' + 1 = '01', etc.

Fix:
- SQL: cast COUNT(*) to ::int so pg returns a number from the start
- Frontend: wrap both increment sites in Number() as a defensive guard

+++ kanecta-app-community-hub/community-hub-api/routes/discussions.js
@@ -47,7 +47,7 @@ router.get("/threads/:threadId/messages", requireAuth, canAccess, async (req, re
   try {
     const { rows } = await pool.query(
       `SELECT id, thread_id, user_id, user_name, content, created_at, edited_at, deleted_at,
-              (SELECT COUNT(*) FROM discussions_messages r WHERE r.parent_message_id = m.id) AS reply_count
+              (SELECT COUNT(*) FROM discussions_messages r WHERE r.parent_message_id = m.id)::int AS reply_count
        FROM discussions_messages m
        WHERE thread_id = $1
          AND parent_message_id IS NULL
